move configuration of several Okta objects to the account subproject
