+++ app/src/main/kotlin/org/cru/godtools/dagger/OktaModule.kt
@@ -1,14 +1,8 @@
 package org.cru.godtools.dagger
 
 import android.content.Context
-import android.os.Build
-import com.okta.authfoundation.AuthFoundationDefaults
 import com.okta.authfoundation.client.OidcClient
-import com.okta.authfoundation.client.OidcConfiguration
-import com.okta.authfoundation.client.SharedPreferencesCache
 import com.okta.authfoundation.credential.CredentialDataSource
-import com.okta.authfoundation.credential.CredentialDataSource.Companion.createCredentialDataSource
-import com.okta.authfoundation.credential.TokenStorage
 import com.okta.authfoundationbootstrap.CredentialBootstrap
 import com.okta.legacytokenmigration.LegacyTokenMigration
 import com.okta.oidc.OIDCConfig
@@ -18,6 +12,7 @@ import com.okta.oidc.storage.SharedPreferenceStorage
 import com.okta.webauthenticationui.WebAuthenticationClient.Companion.createWebAuthenticationClient
 import dagger.Module
 import dagger.Provides
+import dagger.Reusable
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
@@ -25,17 +20,13 @@ import javax.inject.Singleton
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import okhttp3.HttpUrl.Companion.toHttpUrl
-import okhttp3.OkHttpClient
-import org.ccci.gto.android.common.okta.authfoundation.credential.ChangeAwareTokenStorage.Companion.makeChangeAware
-import org.ccci.gto.android.common.okta.authfoundation.credential.SharedPreferencesTokenStorage
-import org.ccci.gto.android.common.okta.authfoundation.credential.migrateTo
-import org.ccci.gto.android.common.okta.authfoundation.enableClockCompat
-import org.ccci.gto.android.common.okta.datastore.DataStoreTokenStorage
 import org.ccci.gto.android.common.okta.oidc.storage.security.NoopEncryptionManager
 import org.ccci.gto.android.common.okta.oidc.storage.security.createDefaultEncryptionManager
+import org.cru.godtools.BuildConfig
 import org.cru.godtools.BuildConfig.OKTA_AUTH_SCHEME
 import org.cru.godtools.BuildConfig.OKTA_CLIENT_ID
 import org.cru.godtools.BuildConfig.OKTA_DISCOVERY_URI
+import org.cru.godtools.account.provider.okta.OktaBuildConfig
 import timber.log.Timber
 
 private const val OKTA_SCOPE = "openid profile email offline_access"
@@ -46,39 +37,11 @@ object OktaModule {
     private const val TAG = "OktaModule"
 
     @Provides
-    @Singleton
-    fun oidcClient(@ApplicationContext context: Context, okhttp: OkHttpClient): OidcClient {
-        AuthFoundationDefaults.enableClockCompat()
-        AuthFoundationDefaults.cache = SharedPreferencesCache.create(context)
-        AuthFoundationDefaults.okHttpClientFactory = { okhttp }
-        val config = OidcConfiguration(OKTA_CLIENT_ID, OKTA_SCOPE)
-        return OidcClient.createFromDiscoveryUrl(
-            config,
-            "$OKTA_DISCOVERY_URI/.well-known/openid-configuration".toHttpUrl()
-        )
-    }
-
-    @Provides
-    @Singleton
-    fun OidcClient.oktaTokenStorage(
-        @ApplicationContext context: Context,
-        coroutineScope: CoroutineScope
-    ): TokenStorage = DataStoreTokenStorage(context)
-        .let { dataStore ->
-            when {
-                Build.VERSION.SDK_INT < Build.VERSION_CODES.M -> dataStore
-                else -> SharedPreferencesTokenStorage(this, context)
-                    .makeChangeAware()
-                    // migrate tokens if the user had previously used the DataStoreTokenStorage
-                    // TODO: this may race with loading tokens in the CredentialDataSource.
-                    //       We may want to create a "MergedTokenStorage" class instead to facilitate token migrations.
-                    .also { prefs -> coroutineScope.launch { dataStore.migrateTo(prefs) } }
-            }
-        }
-
-    @Provides
-    @Singleton
-    fun OidcClient.oktaCredentialDataSource(storage: TokenStorage) = createCredentialDataSource(storage)
+    @Reusable
+    fun oktaBuildConfig() = OktaBuildConfig(
+        clientId = BuildConfig.OKTA_CLIENT_ID,
+        discoveryUrl = "${BuildConfig.OKTA_DISCOVERY_URI}/.well-known/openid-configuration".toHttpUrl()
+    )
 
     @Provides
     @Singleton
