+++ library/account/src/main/kotlin/org/cru/godtools/account/provider/okta/OktaBuildConfig.kt
@@ -1,0 +1,8 @@
+package org.cru.godtools.account.provider.okta
+
+import okhttp3.HttpUrl
+
+class OktaBuildConfig(
+    val clientId: String,
+    val discoveryUrl: HttpUrl
+)
